+++ NEWS.md
@@ -1,11 +1,19 @@
 # mixvlmc (development version)
 
+## Breaking changes
+* the `mode` parameter of `cutoff.vlmc()` has been renamed to `scale`
+* the `"auto"` type has been removed from `contexts()` variants 
+
 ## New features
 * new functions `predict.vlmc()` and `predict.covlmc()` can be used to make one 
   step ahead predictions of a time series based on a (CO)VLMC model (issue #46).
   Those function are documented in  a new vignette (`vignette("prediction")`)
 * `contexts()` can now report the positions of each context in the original time
   series
+* nodes of context trees can be extracted individually as `ctx_node` objects 
+  using the `find_sequence()` function. A collection of new functions can be
+  used to manipulate the nodes and gain fine grain information on the 
+  corresponding sequences (issue #50). 
 * the log likelihood calculations performed by `logLik.vlmc()`, `logLik.covlmc()`
   `loglikelihood()` and `loglikelihood.covlmc()` have been revised, expanded to 
   include three possible definitions of the likelihood function, and documented